Fix test failures on OSX.
Angle test for A-B-A angle (which should be a true 0) was failing
with unacceptable amounts of error. arm64 uses a fused multiply-add
instruction here and floating point error accumulates pretty badly.

Use Kahan's method for difference of products to cancel out the
error, with needed intermediate values at double precision.